Style: change folder and file to MVC design
