+++ public/index.html
@@ -16,6 +16,6 @@
   <body>
     <canvas id="gameCanvas" width="800" height="600"></canvas>
     <script src="/socket.io/socket.io.js"></script>
-    <script src="./game.js"></script>
+    <script src="js/client.js"></script>
   </body>
 </html>
